+++ backend/Mgtt.ECom/src/Mgtt.ECom.Domain/ShoppingCart/ICartService.cs
@@ -2,6 +2,7 @@ namespace Mgtt.ECom.Domain.ShoppingCart;
 
 public interface ICartService
 {
+    Task<IEnumerable<Cart>?> GetAllCarts();
     Task<Cart?> GetCartById(Guid cartId);
     Task<Cart?> CreateCart(Cart cart);
     Task<Cart?> UpdateCart(Cart cart);